refactor: externalisation du scenario Puppeteer dans `default-puppeteer.ts`
